Support foxpro 2 (#11)
* Mid step rename

* Rename to uppercase.

* Add support for FoxPro 2.x.

- Fixed how memo indices are calculated.
- Fixed not returning all memo records correctly when iterating.
- Updated `FirstIndex` in `Memo.cs` to use `GetBlockCount`.
- Refined `Add`, `Get`, and `Set` methods in `Memo` for better functionality.
- Cleaned up unnecessary content files in `DBase.Tests.csproj`.

* Refactor binary data handling in Dbf and Memo classes

Updated the Dbf class to replace MemoryMarshal with BinaryPrimitives for improved performance and clarity in reading and writing binary data types.

In the Memo class, added a new Set8B method for structured writing of multiple MemoRecord instances and introduced SetFP for handling a different record format. Removed the previous Set8B implementation to streamline functionality.